:recycle: type변경
ㅇㅇ

+++ src/main/java/com/g3c1/temi/domain/purchase/presentation/dto/response/PurchasedFoodListResponse.java
@@ -11,7 +11,7 @@ import java.util.List;
 @RequiredArgsConstructor
 public class PurchasedFoodListResponse {
     private final Long seatId;
-    private final int seatNumber;
+    private final Long seatNumber;
     private final List<FoodInfo> foodInfoList;
 
     @Getter @Builder
